Rewrite the remainder of ImageUtil in such a way that it is friendly to asynchronous downloads and conversion to common libraries.
Up to N image loaders are fired at the end of the frame and are checked at the beginning of each frame. If the loader is completed, more are fired so the total is always n downloads in flight. The C# version of this runs synchronously but it is still checked for completion in the asynchronous manner. This enables optimal conversion to JavaScript and other platforms.

Hooray!

------------------------------
Imported from Google Code.
Original SVN rev#: 197