Introduce build file variables that can be set individually per project. Use the following syntax in your build file:
<var type="boolean">
  <id>meow</id>
  <value>true</value>
</var>

change <value> to <env> to make it use an environment variable instead of a literal value.

In the Crayon code, use the following syntax:

$var['meow']

These are compile time dictionaries so they must be dereferenced with a hard coded string constant at compile time. This allows exclusion of code blocks at compile time so that the code doesn't get included in the final build.

As much as I'd like to live in a world where one code base will work everywhere in exactly the same way (my mantra thus far) sometimes it just doesn't work that way due to inherent differences in the platform. For example, if you want to exclude an "exit" menu option for JavaScript targets, that makes sense, and this is how you can do that easily from the build file.

------------------------------
Imported from Google Code.
Original SVN rev#: 279