Refactor how the stack and asynchronous execution works. This breaks all platforms except C#/OpenTK. Fixes coming soon.
Begin the process of obliterating all the weird tick-registered stuff. Now that the types of programs that can be made with Crayon is drastically expanding, this hack was causing a lot of pain.
The new ExecutionContext stuff opens the doors for some interesting stuff. For example, a blocking UI window that natively pauses execution while being shown (but the platform-independent Crayon-lib doesn't treat it that way or vice versa) which also fires handlers on UI events such as mouseClick. It also allows for asynchronous workers to actually be asynchronous and get rid of the async message pump.