Mostly destroy everything. Nothing works now. Especially GFX. That especially doesn't work. I've created a library called ImageResources. This contains a few useful classes: - ImageLoader - ImageSheetLoader (untested) - NativeImageData - stores actual bitmap data. Supports locking of resources for pixel editing OR for GL texture loading - ImageResource - the new "Image" object exposed to the user. Does not do anything other than contain a NativeImageData reference (although not exposed directly to the user) and has a width and height.
ImageLoader supports both synchronous and asynchronous loading. If the underlying platform does not support one of these (e.g. PHP being synchronous only or JavaScript being asynchronous only) the loader will "fake it" so that the end-user code will appear to behave mostly the same.

The ImageSheet code in the compiler has been drastically cleaned up and modified to be more flexible and also support JPEGs within the same logical sheet (whereas they were required to be PNGs before).

I've also deleted all the structs and program data and most of the other code in the interpreter that is GFX-releated. It'll be ImageResources and GFX's responsibility to implement these.
ImageSheets are completely rewritten and the loading logic is handled in embed code in ImageResources.
All references to the word "spritesheet" have been removed. They're all ImageSheets now.