Re-implementation of a good chunk of the GFX library.
Not tested, but hey, it compiles!
Lots of dead code removed.
Lots of live code removed that deserves to be dead.
Now that this compiles, I'll be going back and fixing up the image sheet loading.
Once that works, I'll go through and update the SquishySquid demo project and then actually test this for real.