Add ImageEncoder support for Java. Fixed bug where non-image-sheet images were not being copied to output.
Noticed an issue where Java was generating JPEGs with corrupted colors if the input was a PNG with an alpha channel (even if there were no alpha pixels).
Although I fixed this for Java, ideally there should be a well-defined behavior here. Probably setting the fallback background color in the JpegEncoder class itself with some default color already set.