Library method translators should be able to have a platform-independent-but-language-specific default translators for things that are generic across the language.
For example, this commit converts the JSON library to such a model. Parsing JSON is the same whether you are doing it in a C# desktop OpenTK project or a C# Xamarin iOS project.

Next goal is to migrate a lot of the SystemFunctionTranslations to the Core library instead of the mess of abstract classes that exist now. This is part of a larger effort to eliminate code in favor of config files.