Port VM to Ruby
This adds a project target type called game-ruby-gosu. While this does not fully implement the required game libraries, projects that only use basic VM features can now run on basic Ruby. There's enough work to do that this will likely not make it into the 0.2.0 release as a documented feature, but is standalone and stable enough to not need to keep it in a branch anymore. 

There's also a bug fix for passing a build file path that isn't in the current directory (potentially a regression from the recent multi-<source> stuff that was recently added.