Introducing Pastel.
Currently the Crayon VM is written in a hacked up version of itself. Crayon, being loosely typed, is not the best at this and relies heavily on type annotations for both declarations and dereferencing values. This has been relatively painful and the tipping point to swap this out with something else was reached while trying to port to C.

Pastel is staticly typed and heavily influence by C# and is only used internal for the translation of the VM and libraries.

Furthermore, splitting this off into its own language will allow me to do major cleanup of the user-facing Crayon language itself as one-off hacks like structs and annotations can be taken out entirely.
There will be quite a lot of code duplication, but it's probably going to be orders of magnitude cleaner and safer this way as a siloed language.