Merge pastel to master
Well, it looks like I merged from pastel branch to master branch at some point accidentally and so the benefit of having a separate branch vs getting a stable snapshot from sometime in the past is no longer there. So I'll go ahead and flatten this and then split them back apart if there's a need.