Rename platform projects.
Eventually the goal is to completely de-game all project types. The fact that these are games should be purely be a factor of which libraries are imported, and even then, it should be referred to as a graphical app (e.g. GL) as opposed to a game. This is all to eventually allow a 100% generic standalone VM.