Fix Python which was crashing on exit (which wasn't very noticeable at first)
When a game closes, it was looking for the active execution context to suspend. However, because Python blocks while running the game loop, it actually was cleaning up the ExecutionContext before halting, and then the calling code wasn't able to find the instance, so snapshot a pointer to that ExecutionContext before invoking the screen. This should have no effect on other platforms as the reference should be the same, but Python won't crash now.

Also, cleaned up the Game library old translated .cry files and transferred comments over to the .pst files.