Implement Chrome Web App project exporter
This includes a lot of additions to build file properties such as default window size, project description, and version. These are needed for the manifest.json in the exported project.

Also fix some oddness with the icon file, which seems to have regressed for all projects. This will require more work to get fixed for other platforms.