Python projects -> prevent crash when loading images before window is open
Unfortunately there's no way to do this in PyGame, so sadly open a dummy window.
Successive calls to pygame.display.set_mode() will simply change the window size to the proper size.
It is also unlikely that the dummy window will be up for long.

Modify Asteroids code to trigger this use case.