Fix JS font loading when loading the page for the first time in a tab.
This fixes the issue where the text renderer was falling back to the default font but only the first time you load the page in a tab. If you pushed refresh in the same tab, it would work correctly.
This fix is applicable and tested for Chrome and Firefox. Internet Explorer seems to be broken outright, which will require a wider investigation.

The fix is to actually apply the font to a dummy span and yield to the UI thread for 1 millisecond (using Core.sleep).

Fix bug where compile-time Pastel constant of IS_JAVASCRIPT was not longer working.