Unroll the library portion of stack traces on release builds of the VM.
This feels a little cheesy but for core libraries and beginning users, it feels like an okay compromise for now.
Eventually solutions to make this feel more correct is to perhaps include an exception unroll default behavior in the library manifest, where core library exceptions always unroll back to user code, and custom libraries will not unroll by default. Also allowing a compilation flag or build file arg to override this behavior.