Move some op implementations into their own functions
This will reduce the function size of the main interpreter function which is mostly inlined code for speed.
This was never a problem before but it seems Android Java has a built-in limit for number of compiled ops (which is 2^14=16384 ops).
This brings the total number of Android-Java compiled ops in interpret down to 17232 from 17830, so there's still some more optimizations to go.

The functions that were refactored here are metadata initialization stuff that has no critical performance need to be inlined.