Fix execution suspending/halting/continuing issues with AWT based UI.
Execution context should stop when window.show() is called and resume when the window is closed.
Also ensure that the AWT-based projects end when the process comes to an end. There's some oddities with AWT where the process will stay active even when all windows are closed and the main function completes.
Unfortunately fixing this required a lot of changes to check if execuction context 0 finished cleanly and then forcefully call System.Exit(). This isn't implemented for some platforms since it works correctly for most.