Android platform experiment
A series of experiments and fixes added to the Android platform. Since there were problems with the OpenGL renderer that were likely to take a while to fix, I dabbled with an experiment to use the JavaScript renderer from the already-implemented JavaScript platform and stick it in a webview, while the rest of the VM remained as a Java-implemented VM. It's kind of wonky, and in the end while doing this I realized that the way the VM is currently implemented has serious performance problems that are specific to Java on Android. General desktop-JavaScript, aside from GC hiccups every few seconds, is a much better starting-off point than Java for a longer term Android strategy (tested by simply looking at the Crayon demos on the website via the regular browser on the phone), and so I'm marking this platform with an "experimental-" prefix and will turn my focus towards a javascript-app-android approach. There were a few interesting things I want to leave here for reference, although I'll likely delete this directory after the next version branch.