Perform the grunt work of upcoming refactoring
I'll need to pass around some context to the transpilers. Because the string buffer is passed everywhere and is really just a context object of sorts, I've changed its type.
Remove any remaining weird string munging from the transpilers and push it all into the platform implementation.
Fix a few bugs that slipped in in recent changes
- the PlatformTranslationHelper on C# not having the correct namespace
- the library functions code snippet arg replacer thing had an out of bounds exception.