Fix problems with packaging the GameGifCap library. This will unbreak the Asteroids demo when running directly
GameGifCap library was using a hardcoded reference to the GameWindow, which would only work when exported as a single binary. When exporting, this cross-library reference would break building the libraries. The solution is to apply a reference of the GameWindow to the Crayon Game.GameWindow object, and then pass that as a plain C# object to the GameGifCap's native screenCap function. The screenCap function uses reflection to invoke the ScreenCapture() method.
Additionally, add a reference to the System.Drawing .NET library to the GameGifCap DLL.

Also delete some method definitions for no-longer-supported platforms in the Game library.

Modify the release.py script to include the GameGifCap Library in releases.