If the nth image draw command was an image generated by Graphics2DText where n is a power of 2 greater than or equal to 64 on a platform based on pixel-buffer images rather than OpenGL, then this would result in a runtime null reference error.
The image buffer lazily expands to accommodate more images to send to the render pipeline by doubling it. It starts out at 64, I believe. Some lines of code were nested incorrectly into the value copy loop, and were wiping out the rest of the images beyond the first image, and was causing a null reference error in the render pipeline.