Fix screen aspect ratio problems in Android.
The width always seems to be correct whereas the value of the screen height from scrollHeight always seems random. This fix is a little hacky to avoid this by letting the Java side figure out what the screen aspect ratio is and then computing the screen height as a function of that and the (stable) width.

If someone would like to explain why this is and a better way to do this in the form of a pull request, I'd be very happy.