This makes the Node+Owner relationship a little more consistent
For example, field definitions, while a top-level entity, were not considered the owner of the expressions for the default value for the fields. The class definition was.
Even constructor code had the ClassDefinition as an owner, not the constructor.

Now anything that is construably a member of any top level entity, is. Furthermore, the Owner property is now a Node type instead of a TopLevelConstruct because Lambdas need to be owners now.