Implement Crayon Native Interface (CNI)
Functions can be registered in the manifest.json file and called directly by Crayon code.

It is up to the specific platform code to register a method that takes a name and returns a function pointer and does the necessary magic to find that method in the library.
This is a little complicated, but luckily it is only up to the platform provider (me) to implement this. For the library provider (advanced users), they only need to provide a function in the library native code that has the name that's registered in the manifest, which is much easier.

This also paves the way to eliminate Pastel as a requirement.

FYI: a function pointer is:
* Python -> a regular function pointer (natively supported)
* JavaScript -> a regular function pointer (natively supported)
* C# -> System.Reflection.MethodInfo
* Java -> java.lang.reflect.Method