Undo all the Pipeline-Worker stuff
The same pipeline and workers still exist logically, but as pure C# code.
The driving motivations of creating this architecture was two-fold as I can remember it:
- an excuse to refactor things cleanly in clearly readable pipelines
- isolating responsibilities into well-contained workers that could be individually migrated to an out-of-process worker written in C/C++ so that I could eventually migrate the compiler away from C# in a non-monolithic way. Since migrating directly to C/C++ is no longer on the road map, this is no longer useful.

One of the unfortunate side effects of this is that the performance markers were all in the pipeline interpreter and so I'll have to investigate other options if I want performance metrics again.