The responsibility of what the platform should do when an interpreter finishes is now a bit more clear.
This returns the InterpreterResult from all methods that run the interpreter and the harness that invoked it is now more responsible for what to do with that result.
Furthermore, SLEEP is now purely a harness-side concept and the Pastel implementation of both Thread.Sleep AND EnqueueVmSuspend can now be removed.
This is only implemented in C# and Java. Python and JS are currently broken (although Python was already broken).
This also removes SUSPEND_WITH_LOCK which was confusing and introduces SLEEP as an interrupt type.