Completely change how dictionaries work.
Previously, a dictionary would contain 4 native dictionaries:
- string to Value for the key
- string to Value for the value
- int to Value for the key
- int to Value for the value

Now there are just two dictionaries that go from native int or string to an integer index value.
Additionally, there are two parallel lists for the Values for the keys, and Values for the values. The lookup dictionaries point to an index in these lists.
This makes accessing the keys and values much more streamlined as getting the index from the raw key is the only step that is different between different key types.
Additionally, it makes the iteration order of dictionary keys both easier and consistent between platforms.