In dynamic scopes, access a private field from an object should result in an error.
This commit reads the access modifier data from the bytecode and puts it in the class metadata table and uses it at runtime when dereferencing a field from an instance.
Assigning a field still needs to be implemented. Currently there are only private checks. internal and protected still need to be implemented.

Additionally this creates a in-bytecode cache of the most recently accessed class type from a particulr DOT_DEREF instruction. I anticipate this cache to hit 99% of the time thus eliminating the need for doing a dictionary lookup of the localized nameID to figure out the field/function member ID in most cases. There was also an extraneous dictionary lookup as well that wasn't removed from before that was also removed. Even when protected/internal checks are added, this should still be an overall performance boost.