Add a new op to verify statically typed function arguments.
This op is currently un-implemented, but uses the same arguments as cast and is injected between the arguments and body of the code of a function.

In the future:
When a function pointer is invoked with no type information, a boolean flag will be set on the stack frame indicating that the type information needs to be verified. If the flag is false, the op is skipped, otherwise, the type verifier will run like a cast and if an error is encountered, the stack frame will be popped once and an exception thrown.