ListImpl uses an array or list implementation for its items.
For languages where there is no distinction between an array or list (JS and Python) this doesn't make much of a difference.
For languages where there is an expandable list type (C# and Java), this basically inlines the extraneous logic for that. This had a pretty good performance boost in Java when there used to be a FastList hack in Pastel designed just for the ArrayList<Value> types (which was later removed because Pastel needs to be Crayon-free). C# has a similar implementation for System.Collections.Generic.Lists where the List class is simply a native array with unoptimized C# code performing the List logic.