Add a new flag-accessible compilation pipeline that only performs compilation without generating any files or resources.
All errors with location information are returned to STDOUT in an easily parseable JSON format.
This can be used by an IDE.

I've also made it so that any flags that are prefixed with "-CR:" will be trimmed and sent only to the compiler process (as opposed to being misinterpreted as end-user program command line args).
Other arguments ought to be converted to use this format (although I'll wait on requiring this until there's an IDE as this will be kind of a pain for normal usage).