Implement the STEP-OVER debugger command.
This (ab?)uses the interrupt system to achieve step over.
A new struct has been created called DebugStepTracker. This struct is applied to the current stack frame when the user is stepping through code.
Additionally, a persistent interrupt of type DEBUG_STEP is generated and hasInterrupt is always set to true during a step over. This interrupt also has a reference to the same DebugStepTracker instance.
When the stack frame changes, the execution context's activeInterrupt and the hasInterrupt boolean are updated accordingly.

For function calls, the DebugStepTracker is set to null and when the hasInterrupt is triggered, the interrupt is then cleared because it no longer matches the stack frame instance. As such, there is no special code in the call function code to achieve this as it is handled entirely by interrupt code when it gets triggered but sees no matching DebugStepTracker in the stack frame.

For returns and exception bubbling, the new stack frame is explicitly checked to see if there is a DebugStepTracker instance. If so, hasInterrupt and activeInterrupt are updated accordingly.

Unrelated bug fix that was not practical to split from this change:
- Fixed bug where a breakpoint that gets hit will immediately get hit upon resume if you don't clear it first. It should be ignored when it's the first PC encountered after resume since presumably it's the breakpoint that you're resuming from.