Remove interpreter source from the compiler
This deletes the Interpreter.csproj project and puts all the source in Interpreter/source. Because this is still needed (for now) when you export a project to a platform, it also needs to be included in the final release package. 

Unrelatedly, this also fixes the annoying disk IO race crash in the release.py that always happened on the 2nd time you run the script without deleting the previous release files. (simply add a sleep(0.1))