Introduce a partial implementation of namespaced scopes for Pastel.
This breaks things, but hey, that's what branches are for!

PastelCompilers now have a collection of shared scopes that also have a namespaced prefix associated with them. This is hardcoded as "VM" for now. The code changes for this aren't implemented yet, as there needs to be a significant rewrite with how PTypes are parsed and resolved after this change is committed.