Do a Resolution pass in Pastel to finalize struct references in PType for function and struct signatures.
Pass the parser context around for all CheckAssignment/IsIdentical invocations so that FinalizeType can be used. Because the value of finalized struct is cached, it's okay to pass the current pastel compiler instance instead of the one referenced by a namespace when doing these comparisons (in cases where the type comes from an argument/return type from a function definition defined in a different pastel compiler scope).