Remove (almost) all references of the word "Native" in the Pastel project.
This was used ambiguously to refer to either the Core functions or the Extensible functions and was confusing.
Now everything is either a Core function or an Extensible function.
The actual Pastel code itself still uses "Native.FunctionName" for extensible functions and so that will be removed in a separate commit.