Refactor Audio and Gamepad libraries
The native code in the Audio and Gamepad libraries mostly just call into helper functions in the Game library. By putting the invocations of the native code directly in the Game library's embedded code and just call into that from Audio and Gamepad, which are now just embed-only libraries, I can remove the last cross-project dependencies for when code is exported to C# and remove support for this feature altogether.