Rename library resources/ directory to native/
Each platform specific instructions file is now called manifest.txt and located in a directory that has the name of the platform. Files that were in these directories are now in some sort of directory that describes its contents (typically "source", but also "dlls", and soon there'll be a "generated" directory as well).

Deleted a stray manifest.txt from the Core library root from the days of yore.