Create a dummy project called AcrylicConvert that loads the C#-written compiler as though it is Acrylic code.
Create a flag to indicate that the code should be parsed in C#-compatibility mode. This will use a new platform type called AcrylicConversion that will transpile the parse tree to true Acrylic rather than serialize a VM. 

Currently the Acrylic parser doesn't support true C#. This will be added little-by-little in this branch when this new flag is active.