Implement the basic Core functions for a Dispatcher.
While this doesn't really do anything interesting on its own, it allows for UI frameworks to register themselves as a handler for triggering timed callbacks.

It also includes a function called flush() that allows you to trigger anything that needs to be triggered if you are in a UI-free situation and are just running a basic command line process with no UI-based dispatcher.