Fix a bunch of random Java Pastel generation errors (caught by unit tests! yay!)
IS_JAVA was not being set as a Pastel flag and so string equality checks were using == instead of .equals()
Fix the extraneous semicolon in the Java import headers in generated code.
Re-generate all the Pastel stuff for Java for safe measure.
There was a place (likely from the Crayon-to-Pastel transition era) where hard coded integers were being used instead of the Types enum.
Fix a typo and an outdated class name in JsonParser.java