Update "default-title" build property to just "title".
Added a bit more consistency to the naming conventions of PROJECT_ID/PROJECT_NAME/DEFUALT_TITLE/TITLE/PROJECT_TITLE. There is now no such thing as a "default title" or a "project name". "project title" and "title" are interchangeable terms that refer to the readable formatted name, and "project id" and "id" always refer to the alphanumeric ID.

If "default-title" is encountered, the compiler will give an error message instructing the user to update their build file.

Added a title to a lot of the demo files since they all seemed to be missing it.

Updated docs.