Do not run the next frame of the JavaScript Game engine after initializing the screen until you've given the VM a chance to suspend itself.
This is done with a simple setTimeout(nextFrameCallback, 0) instead of calling it directly.
The way it was happening before was causing two VM invocations to be nested within each other and the outer one would suspend itself after the inner one was completed, which caused it to revert back to the state it was when finishing the original outer invocation and madness ensued.