If a variable-like expression is not declared within a const definition, this should throw a reasonable error. Currently crashing.
This was because variables are not allowed in const definitions and so the variable resolution phase was being skipped for constants. However, this resolver critically determined whether a variable-like expression was not previously declared. As there is no way to declare a variable, this will not pick up any variable definitions, but still do the important step of throwing errors.

Test added.

This fixes: https://github.com/blakeohare/crayon/issues/262