Move all byte code generation from the exporter assembly into the parser assembly.
The parser and byte code compiler has no reason to be separate. They both make use of the ParseTree which can now be made internal to just the parser assembly.

The exporter should only take finalized resources, including the final byte code, and assemble them into end-user files.

"Compiler" is a better name for the parser assembly, but I'm not going to rename it yet.