Moving development of U3 from its own repo directly to Crayon. This is still very much a work in progress but it won't interfere with any existing features.
Eventual goal of U3: U3 (Universal UI Utility) is a chunk of JavaScript code that can render both traditional and game-based UI via a simple 1-dimensional stream-based protocol. This JavaScript code is wrapped in an Electron window that can work without dependencies on Windows/Mac/Linux and (eventually) can be wrapped in a WebView for iOS and Android, and since it's JavaScript, can work on the web trivially. This is similar to the internal implementation of the Nori library (wrapped in a WinForms window) and also the JS implementation of the Game library. My goal is to migrate both of those libraries to use U3.

Benefits:
- Remove dependence on OpenTK (which has proven flakey when on non-Windows platforms)
- Remove dependence on WinForms (which has blocked migration from .NET Standard to .NET Core. It also doesn't work at all on non-Windows platforms)
- Remove inconsistencies between platforms (.NET vs JavaScript implementations can look different, especially when text is involved)
- Remove internal image atlasing which was critical for performance in OpenTK and has held back adding interesting features to the ImageResources library