U3 Named Pipe should accept strings of any size without breaking them up.
This seems to be impossible and so instead of handing off the strings directly to the render process, a length + payload format has been created. Strings are sent with the desired length preprended to them followed by a delimited followed by that many bytes of payload. `namedbigpipeserver` only calls the onData callback when it receives that many bytes.

The protocol only uses simple lower ASCII characters, so bytes vs character count isn't an issue.