Serialize out a secondary image manifest that is currently ignored.
This new format does not support user-named image sheets (it automatically groups things by proximity by filename and image loaders will simply take in file path prefixes) and is smarter in how it packs images together. Additionally the format supports the notion of images as library resources, not just images within the user-written code.

There is an annoying '2' in a lot of these names to avoid collision with the current Image Resource. This will be removed once fully migrated and the old code can be safely removed.