Implement Bitmap.fromBytes(byteList) and Bitmap.fromBase64(b64String) for both C# and JavaScript platforms.
Previously was just Bitmap.fromBytes() for the C# platform only.

C# naturally uses a byte array and JavaScript naturally uses base64, and so these both pass their natively-preferred formats to the native code.

JavaScript requires loading to be async, and so the core function implementation runs synchronously for C# and asynchronously for JavaScript and it's all very confusing.