U3Graphics2D now supports drawing images.
Fix race condition where it was possible to send Nori events before the window was ready and these were getting dropped. This was already dropping Game draw events, but I didn't notice before until I started sending image load data on the first event only.

There's a bug where the images are skipped if they aren't loaded, and so images aren't shown on the first frame that they are used. This could be pretty bad, and so I'll have to re-think the API here. For now, I'm letting this bug be to unblock further Game library conversion.