Rewrite the Matrices library to be pure Crayon without any CNI.
The main use case for this is the Asteroids demo which requires the more of the Game libraries to be completed to be tested, so this isn't fully tested. Mainly my concerns are whether the API matches before and the way this was constructed was a little unclear. Like .toVector().