Fix various bugs in the JavaScript implementation of pixel edit session. Added more friendly-named wrapper functions for image flip.
Some of this was simply not implemented fully and still had bits of the old partial implementation in it that used imageData more extensively.
GetPixel was appending to the end of a list as opposed to clearing it out and was not being passed clean lists.
Fixed regression in the test_server.py script I probably introduced with the last favicon change.