Convert FileIOCommon to use Core Functions instead of CNI.
This concludes converting the last of the libraries that are actively being packaged by the release script to use Core functions. However, the standalone VM and release script are expecting there to be at least one library present and so it is no longer building correctly. For this, I have created a new dummy library called "Dummy" that has a CNI method registered in its manifest so that the VM/releaser still works.

I'll address this separately. But for now, FileIOCommon is DONE.