U3: u3init message passes along a base64 encoding of an icon file that is used by the Electron BrowserWindow constructor.
The icon is a temporary image created by U3Window that's just a blue square in a red square.
There seems to be some problems with the IconEncoder and the actual icon bytes don't seem to be correct. I swapped out the bytes with an actual .ico file and the rest of this pipeline seems to work correctly.

Since an actual file is required by BrowserWindow, the icon bytes are written to the user's temp directory.