Rewrite the Audio library.
This is only a partial implementation that supports loading raw samples/channels of audio and playing it. However this implements the new audio pipeline end-to-end, so it's a good spot to do a commit.

The basic Audio is implemented in U3 and so it can be used in both Games and Nori based projects.