C# VM image stuff: Parentheses were not grouped correctly and causing colors in translucent pixels to become garbled.
Cause: There are two valid constructors for Rgba32 structs: one that takes in bytes and one that takes in floats. The cast to byte was not being applied to the final division by an integer, and so the expressions being passed in as args were actually ints, which get converted into floats when C# decides which constructor signature is compatible with those types. The float constructor maxes out all values at 1.0 so basically all channels of translucent pixels were being treated as 0 or 255 with nothing inbetween.