Add a field to Buttons to enable/disable them.
Disabled = true is the usual intent, but having a positive tense property for boolean fields is better for consistency, so I've added both, where they're just aliases to the same property. Because false is the default value when a property isn't, disabled is the real underlying property sent to U3, and enabled just inverts the value of disabled.