You can now query the rendered size of an element.
This adds a new property and event type called render query. The render query property on element is a pseudo property that invokes some sort of method that then triggers the render query response event. This is a handy way to implement arbitrary render-side functionality that can return results to the VM-side without creating more complexity and message types.

This event and property are located on the root internal Border element of the Frame instance.