Flesh out the codepath to linking up the python language frontend, mostly in the Extension and Disk services.
Python has been added to the programming languages enum. If the language is set to Python, the BuildContext is intercepted after it is generated but before it's sent off to the compiler service and sends the project to the transpiler via the extesion service. This is mostly stubbed out for now.