Split out the properties in CbxBundleView into two classes.
CbxBundleView (now renamed to just CbxBundle since I didn't really stick to the -View naming convention for the JSON-based objects) only contains information that is supposed to be contained in the final CBX Bundle.
Everything else such as build error information and export instructions are now in BuildData which also has a nested reference to CbxBundle.