Exporter is now WaxService-based.
CbxVmBundle references are now out of the exporter and into the main router.
Only the Exporter project has references to the platform projects.
The hierarchy resolution has been simplified in a hacky way. This is all temporary until it transitions into extension-based services.
The main crayon project should not have any references to the Exporter project other than for loading the ExportService. This opens the way to separate the Router service as its own project without exporter dependencies.