Things that use a unique ID in the compiler (such as class ID's) need to be allocated on a per-compilation basis.
These were being allocated by a static int variable that got incremented each time a new ID was needed. This is sometimes sort-of okay (although causes inconsistent results depending on what compilation requests ran first), but certain things are assumed to be allocated from 0 or 1 and if the allocation starts at a big number, it throws everything off (like class ID's)

The compiler service needs to be completely scrubbed of all static caching.